requirements.txt
fastapi
uvicorn[standard]
python-amazon-paapi
ebaysdk
python-dotenv
psycopg2-binary
sqlalchemy
celery
redis
